# docs/openapi/raven-services.yaml
openapi: "3.1.0"
info:
  title: Raven AI — Microservice API
  version: "1.0.0"
  description: |
    Enterprise-grade personal AI assistant.
    Monolith API at :18888, microservices behind gateway at :8000.
servers:
  - url: http://localhost:18888/api
    description: Monolith API
  - url: http://localhost:8000/api/v1
    description: Gateway (microservice mode)
paths:
...
  /tasks:
    get:
      summary: List tasks
      tags: [Task]
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Task list
    post:
      summary: Create task
      tags: [Task]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                goal: { type: string }
      responses:
        "201":
          description: Task created
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...